requirements.txt
flask
pdfminer.six
PyMuPDF
pytesseract
pillow
pandas
matplotlib
scikit-learn
requests

